pyproject.toml
[project]
name = "multi-agents-system"
version = "0.1.0"
description = ""
requires-python = ">=3.12,<3.14"
dependencies = [
    "mesa",
    "pydantic",
    "matplotlib",
    "seaborn",
    "pyyaml>=6.0.3",
    "solara>=1.57.3",
    "networkx>=3.6.1",
    "altair>=6.0.0",
]
